adding image into database and making adding bio and image page
There I add images in data base but this is not a good way to add if I use any type of API then it will be a good way I will use it in version 2 of this application and there I add image and bio adding page it will came when our user don't have any image then this page came but there is a problem that our image is upload in database but now came in our whole application I think due to static adding if I can't fixed it then I will add API in this version and I will add many more things in this version like animation and background images etc as soon as I can before my exams. They are start at 30 Aril 2025 and today is 17 April 2025 so I have very less time to prepare them I will do this application and exam both together Insha'Allah.